Expand utils to update (#12)
Expand RI utils to support modification of the resource.
Using JQ assignment syntax we can leverage the path in RI to modify a given resource.
This PR includes:

jq package (renamed from query) - handles assign and zipAssign operators. zipAssign is needed in case the "flavor" of the RI is with instance ids
Added needed methods to modify the pod spec template, pod spec , metadata, and fargmented pod spec - the caller should use the right funcition based on the variance of the Pod spec path of theirs RI